requirements.txt
anyio==3.6.2
click==8.1.3
DateTime==5.0
fastapi==0.91.0
h11==0.14.0
idna==3.4
numpy==1.24.2
pandas==1.5.3
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sniffio==1.3.0
starlette==0.24.0
typing-extensions==4.4.0
uvicorn==0.20.0
zope.interface==5.5.2
SQLAlchemy
